fast-auth/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.2
h11==0.14.0
idna==3.8
mysql-connector-python==9.0.0
mysqlclient==2.2.4
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
